[BEAM-11593] Move SparkStructuredStreamingRunnerRegistrar to its own package
